requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
httpx>=0.25.0